Agregando funcionalidades al endpoint Playlist

+++ allmusic/src/main/java/ar/edu/unnoba/poo2024/allmusic/repositories/PlaylistRepository.java
@@ -3,5 +3,5 @@ package ar.edu.unnoba.poo2024.allmusic.repositories;
 import ar.edu.unnoba.poo2024.allmusic.entities.Playlist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {
+public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
 }
