+++ allmusic/src/main/java/ar/edu/unnoba/poo2024/allmusic/services/UserServiceImp.java
@@ -23,7 +23,7 @@ public class UserServiceImp implements UserService{
         userRepository.save(user);
     }
 
-    public User getReferenceId(Integer id){
+    public User getReferenceId(Long id){
         return userRepository.getReferenceById(id);
     }
